+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.44
+
+- `title` component option bug fix
+- issue #1: style `scoped` now available
+
 ## 0.0.43
 
 - Option `autoTitle` is now `title` and the default is `false` to disable *title* generation.